+++ CHANGES.md
@@ -1,6 +1,15 @@
 videojs-record changelog
 ========================
 
+3.8.0 - unreleased
+------------------
+
+- Bump required version for videojs-wavesurfer to 2.9.0 or newer
+- Fix `RecordRTC.MediaStreamRecorder` import when using the `timeSlice` option
+- Examples: add `timeSlice` example demonstrating use of `timestamp` event
+- Bump required version for webrtc-adapter to 7.2.8 or newer
+
+
 3.7.1 - 2019/05/30
 ------------------